significant improvements to the makefile - starting each component now stops it if it's already running, so it's now very safe to write make all start_orchestrator_nox. Also cleaned up naming and added some further functionality.
Also, got the feed status changes to work and wrote and tested a python script to test that. Call with: python test_status_change.py cbe1ef798cdbe3ed77d316aed25521059ff1c944, after having changed the status field in the couchdb.

Stopping a feed works in the sense that the erlang processes die and killing off the actual plugin does not cause it to be respawned. However, the erlang process dying doesn't kill off the plugin as that's the subject of another bug: 20601
When a feed is stopped, setting the active status to true and running the python script does indeed force the plugin to be started.

--HG--
branch : bug20587