The merge which is the parent of this rev was bad and broke lots of the python harness. It would seem no testing was done after the merge. I don't know yet whether this was an automatic merge and hg's 3-way merge making _very_ bad decisions, but I am now quite worried about the possibility of further merges going wrong.
I took the diff to the bug20626 parent, edited manually and applied back in reverse. Python harness is back working again.

I also spotted that there was an unnecessary import in the Java Logger, and that the xslt transformer was wrongly catching errors, meaning txns were being committed. Both fixed.