made egress work.
The story:

Every for every output of every pipeline component (PC), there is an exchange.
For wirings that take the output of a PC to a terminal, we create, when the feed starts up, a shovel that binds its own queue to the relevante PC exchange, receives all messages, adds the routing key of the terminal in question, and sends on to the server exchange.

--HG--
branch : bug20673