made list_feeds MUCH more useful.
The first arg is the type to list - i.e. feeds or terminals or servers
The second arg is optional and is 'active' - i.e. only show active things
The third arg is optional and is 'dump' - i.e. dump the full json of the documents
The rest of the args are optional and if they exist they are treated as document ids and used to filter the results (i.e. an entry is only displayed if its _id occurs in the remaining args)