bug20829 Made terminals more flexible by giving them the ability to refer to multiple servers, with different configs. Eg see examples/test/feedshub_status/socket_source_45678.js which specifies the same server twice, with different ports.
Corrected all java servers to cope with this. Scala servers are not yet touched, but still seem to work as it seems they don't inspect per-terminal configuration just at the minute.

Also corrected quite a lot of bugs in the socket_server which had meant that turning terminals off and on for it was as reliable as pissing blindfolded in thewind.

Finally, changed the change_status.py script to make it work for the new terminal structure.