Bug 21481 and bug 21512: Don't try to enter the monitor for publish, since that will be invoked in handlers, which are already inside.  And since servers *won't* usually be publishing in handlers, commit for them on publish, otherwise messages will never enter the system.
There's still a race here: publish for servers /ought/ to enter
the monitor, since publish has to deliver more than one frame and may
result in interleaving frames.  But this would have the same problem
as above for the relay server, since it *is* responding to an input
when it publishes.  So there's a minor danger that either an ack or a
commit on the command input will overlap with a publish on the
output.